requirements.txt
Django==1.8.5
django-cors-headers==1.1.0
djangorestframework==3.2.4
mysqlclient==1.3.6
wheel==0.24.0
